\name{predict.crm}

\alias{predict.crm}

\title{Predict method for CRM fits}

\description{Obtains predictions from a fitted \code{crm} object.}

\usage{
\method{predict}{crm}(object, newdata = NULL, \dots)
}

\arguments{
\item{object}{a fitted object of class "\code{crm}".}
\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted coefficients are used.}
\item{...}{further arguments passed to or from other methods.}
}

\details{...}

\value{
\code{predict.crm} returns a vector of predicted response values.
}

\author{Peter Filzmoser, Sebastiaan Höppner, Iréne Ortner, Sven Serneels and Tim Verdonck}

\references{...}

\seealso{\code{\link{crm}}, \code{\link{predict}}}

\examples{
# Provide an example (TODO).
}
